# openapi.yaml
openapi: "3.2.0"
info:
  title: sing-box hub
  version: 1.3.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    独立的多节点 sing-box 监控管理面板。

    - **数据面**:复用每个 sing-box 节点内置的 Clash API(`experimental.clash_api`)采集
      实时流量、连接(含客户端来源 IP)与内存信息;节点侧无需安装任何组件。
    - **能力**:多节点注册与健康检查、流量历史(分钟/小时/天)、按来源 IP 的流量统计、
      出口 IP 探测、基本控制(断开连接、切换节点选择、切换 mode)、订阅导出
      (sing-box JSON / Clash YAML / base64 分享链接 URI 列表)。
    - **约定**:错误一律 RFC 9457 Problem Details(`application/problem+json`);
      列表游标分页(`cursor` + `page_size`,响应 `next_cursor` 与绝对 URI `links`);
      时间 RFC 3339 UTC;字段 snake_case;数量字段带单位后缀(`_bytes`、`_bps`、`_ms`)。
    - **订阅安全取舍**:`GET /sub/{token}` 为「能力 URL」(capability URL),令牌置于路径而非
      Authorization 头,原因是订阅客户端 App 无法携带自定义请求头。缓解:令牌可随时轮换与吊销、
      端点只读、有限流与访问日志、建议部署侧强制 HTTPS。
    - **节点不可达**:面板到节点的 Clash API 请求失败统一返回 `502`,
      `type: https://sing-hub.dev/probs/node-unreachable`。
    - **加密配置导入**:节点可选 `config_url` + `kce_key`;添加/更新时面板自动拉取
      (支持 KCE1 加密封装)并解析 Clash 代理,全部代理自动导入为面板节点;上游
      拉取/解密失败统一返回 `502`,`type: https://sing-hub.dev/probs/upstream-unavailable`。
    - **账号与会话**:所有面板账号均有管理权限。访问令牌绑定数据库中的会话;
      删除账号、修改密码或用户名、登出会撤销相关会话。刷新令牌只能消费一次。
      升级前的访问令牌需要刷新,有效的旧刷新令牌仍可换取新会话。
    - **公开状态**:`/api/public/*` 无需登录,公开节点名称、标签、出口 IP、版本和运行统计,
      不返回 API 地址、API secret、来源 IP 明细或出站配置。
servers:
  - url: http://localhost:9090
    description: 本地开发默认地址
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      operationId: listUsers
      summary: 管理员账号列表(按 ID 游标分页)
      tags: [users]
      parameters:
        - { $ref: "#/components/parameters/cursor" }
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      operationId: createUser
      summary: 创建管理员账号
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserInput" }
      responses:
        "201":
          description: 已创建
          headers:
            Location: { schema: { type: string, format: uri-reference } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/ContentTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "500": { $ref: "#/components/responses/InternalError" }
  /users/{user_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
    get:
      operationId: getUser
      summary: 读取管理员账号
      tags: [users]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    patch:
      operationId: updateUser
      summary: 原子修改用户名或密码,同时撤销该用户的全部会话
      tags: [users]
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: "#/components/schemas/UserPatch" }
          application/json:
            schema: { $ref: "#/components/schemas/UserPatch" }
      responses:
        "200":
          description: 已更新;空对象不修改账号或撤销会话
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/ContentTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      operationId: deleteUser
      summary: 删除账号及其会话;默认管理员(ID 1)不可删除
      tags: [users]
      responses:
        "204": { description: 已删除 }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录 /session 获取的访问令牌
  parameters:
    cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: 分页游标,来自上一页 next_cursor
...
  responses:
...
    BadRequest:
      description: 请求无法解析
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ContentTooLarge:
      description: 请求体超过大小限制
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    UnsupportedMediaType:
      description: 不支持的请求媒体类型
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InternalError:
      description: 内部错误
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    Unauthorized:
      description: 未认证
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: 资源不存在
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: 冲突(如节点重名)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationError:
      description: 语义校验失败
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    User:
      type: object
      required: [id, username, created_at]
      properties:
        id: { type: integer, minimum: 1 }
        username: { type: string }
        created_at: { type: string, format: date-time }
    UserList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/User" }
        next_cursor: { type: [string, "null"] }
    UserInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 64 }
        password:
          type: string
          minLength: 6
          maxLength: 72
          description: UTF-8 编码后不得超过 72 字节
          writeOnly: true
    UserPatch:
      type: object
      properties:
        username: { type: [string, "null"], minLength: 1, maxLength: 64 }
        password:
          type: [string, "null"]
          minLength: 6
          maxLength: 72
          description: UTF-8 编码后不得超过 72 字节;null 等同于省略
          writeOnly: true
...
    Problem:
      type: object
      description: RFC 9457 Problem Details;type 为稳定绝对 URI,未知类型为 about:blank
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        errors:
          type: array
          description: 校验错误明细,pointer 为 RFC 6901 JSON Pointer
          items:
            type: object
            required: [detail, pointer]
            properties:
              detail: { type: string }
              pointer: { type: string }
      examples:
        - type: https://sing-hub.dev/probs/node-unreachable
          title: Node unreachable
          status: 502
          detail: "dial tcp 192.168.1.10:9090: connect refused"
          instance: /nodes/3fa85f64-5717-4562-b3fc-2c963f66afa6/connections
...